Assume module jars persistent
